Improved reading of CSV files.
* clb/clb/csv.py:

There is now one general function to read all CSV files. The function returns a
dictionary containing data arrays. The arrays can be accessed through their key
specified in the file. An additional array with matplotlib timevalues is
created automatically.

It is possible to only read selected variables which are specified in a list.

* scripts/cloud_height.py: Use newly implemented function read().