Improved reading of scattering coefficients.
* clb/clb/csv.py:

The actual reading of the CSV file in read_scat() is now done using read().
read_scat() is only stacking the arrays for different heights to get a matrix
which is easier to handle.
The function uses regular expressions on variable names to determine which
variables have to be stacked.

The variables stored to the dictionary are now called CLB_MATRIX and CLB_Z
for consistency.

* scripts/cloud_height.py: Adapted new naming conventions.