Run ARTS jobs for all FASCOD atmospheres.
* arts/arts.job:

The job script now uses the environment variable ATMOSPHERE to determine which
FASCOD atmosphere has to be used. If it is unset it defaults to 'tropical'.

The variables ATMPATH and RESULTS_DIR are created depending on the used
atmosphere. Needed result directories are created if not existing.

* arts/atmosphere_batch.py:

Send the arts.job script to the queue for a FASCOD atmospheres.

This script actually sets the ATMOSPHERE environment variable.

* arts/pyrgeoemeter.arts: Use the variable ATMPATH for input atmosphere.

* scripts/radiative_transfer.py: Loop over all atmosphere results.